style: 타입은 따로 파일로 분리하여 관리
