+++ src/components/common/Form/type.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+import { InputProps } from '../Input/type';
+
+export interface InputFormProps extends InputProps {
+  label?: ReactNode;
+  helperText?: string;
+  hasError?: boolean;
+}
